AI4EUEXP-39 Make Hyperlinks possible between catalog entries (and external)
This patch adds the possibility for solution revisions to have
hyperlinks and also reference other solution revisions. CDS upgraded to
version 3.2.0 in this patch. Details as follows.

- Add MLPSolRevHyperlinkMap and MLPSourceRevTargetRevMap
- Cascade the changes needed in client and server as follows
- Add TARGET_PATH in CCDS constants
- Update Rest client with usual methods for the two map tables
- Add repositories for the map tables
- Enable find by revision id in hyperlink repository
- Enable find by hyperlink id, target id, and source id in revision
repository
- Add a Hyperlink service to delete hyperlink with cascading operations
and to delete an orphan hyperlink (not related to any entity). This last
method is to prepare the future where hyperlinks may be associated with
more than just solution revisions.
- Hyperlink controller calls hyperlink service methods to cascade
- Revision controller has methods to get / add / drop hyperlinks,
target revisions, and source revisions
- Solution controller, when deleting a revision, now deletes its
hyperlink and source/target maps. Also deletes orphan hyperlinks if
needed.
- Update client and server tests
- dbConfig files updated with both creations scripts and upgrade scripts
- pom.xml updated to version 3.2.0

Change-Id: I8b2b53bafc29caa1ceb86f4f09020795a43e66c2
Signed-off-by: Shuai Li <shuai.li@cea.fr>